requirement.txt
alembic==1.7.7
bcrypt==3.2.0
cachelib==0.6.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
filelock==3.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
paystackapi==2.1.0
platformdirs==2.5.2
pycparser==2.21
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.34
stripe==3.4.0
urllib3==1.26.9
virtualenv==20.14.1
visitor==0.1.3
Werkzeug==2.1.0
WTForms==3.0.1
